requirements.txt
aiohttp==3.7.2
aiopg==1.0.0
assertpy==1.1
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
idna==2.10
iniconfig==1.1.1
multidict==5.0.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
PyPika==0.44.0
pytest==6.1.2
requests==2.25.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.2
